memory_service/requirements.txt
psycopg2-binary
sqlalchemy
litellm
pydantic
apscheduler
python-dotenv
google-generativeai
prompt_toolkit
